Move exception & diagnostics to API
