+++ palantir-java-format-api/src/main/java/com/palantir/javaformat/java/FormatterException.java
@@ -1,25 +1,26 @@
 /*
- * Copyright 2015 Google Inc.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.palantir.javaformat.java;
 
 import com.google.common.collect.ImmutableList;
-import com.palantir.javaformat.FormatterDiagnostic;
 import java.util.List;
 
 /** Checked exception class for formatter errors. */
-public final class FormatterException extends FormatterExceptionApi {
+public final class FormatterException extends Exception {
 
     private ImmutableList<FormatterDiagnostic> diagnostics;
