+++ palantir-java-format-api/src/main/java/com/palantir/javaformat/java/FormatterService.java
@@ -7,6 +7,5 @@ import java.util.Collection;
 public interface FormatterService {
 
     ImmutableList<Replacement> getFormatReplacements(
-            JavaFormatterOptions options, String text, Collection<Range<Integer>> toRanges)
-            throws FormatterExceptionApi;
+            JavaFormatterOptions options, String text, Collection<Range<Integer>> toRanges) throws FormatterException;
 }
